Get command center pipeline running with agent detections
