+++ mar_bringup/launch/carla_trust_experiments/nodes/cte_agent_passive.launch.py
