+++ rosbag_playback/play_rosbag.sh
@@ -5,7 +5,7 @@ set -e
 ros2 bag play \
     --loop \
     --delay=1 \
-    --rate=0.5 \
+    --rate=0.1 \
     --clock=100 \
     --qos-profile-overrides-path qos_override.yaml \
     '/data/shared/CARLA/rosbags/carla_dataset'
